+++ python/data_structures/hashtable/hashtable.py
@@ -1,7 +1,4 @@
-from linked_list import LinkedList
-
-
-class HashTable:
+class Hashtable:
     def __init__(self, size=1024):
         self.size = size
         self._buckets = [LinkedList()] * size
@@ -32,3 +29,31 @@ class HashTable:
             else:
                 current = current.next
         return False
+
+
+class Node:
+    def __init__(self, data):
+        self.data = data
+        self.next = None
+
+
+class LinkedList:
+    def __init__(self):
+        self.head = None
+
+    def add(self, data):
+        if not self.head:
+            self.head = Node(data)
+        else:
+            current = self.head
+            while current.next:
+                current = current.next
+            current.next = Node(data)
+
+    def display(self):
+        values = []
+        current = self.head
+        while current:
+            values.append(current.data)
+            current = current.next
+        return values
