--- python/data_structures/hashtable/linked_list.py
@@ -1,26 +1,0 @@
-class Node:
-    def __init__(self, data):
-        self.data = data
-        self.next = None
-
-
-class LinkedList:
-    def __init__(self):
-        self.head = None
-
-    def add(self, data):
-        if not self.head:
-            self.head = Node(data)
-        else:
-            current = self.head
-            while current.next:
-                current = current.next
-            current.next = Node(data)
-
-    def display(self):
-        values = []
-        current = self.head
-        while current:
-            values.append(current.data)
-            current = current.next
-        return values
